+++ src/utils/linearop/operator/matmult/spline.cpp
@@ -16,24 +16,24 @@ namespace astroqut
  *  \brief Builds an Abel transform matrix with diagonal radius, without duplicating data or inserting zeros
  *  \param pic_side Width of the square picture
  */
-Matrix<double> Spline::Generate( size_t pic_side )
+Matrix<long double> Spline::Generate( size_t pic_side )
 {
     size_t height = 2*pic_side;
     size_t height_sqrt = (size_t) std::sqrt(height);
 
-    Matrix<double> temp(0.0, height, pic_side);
-    Matrix<double> result(pic_side, pic_side);
-    double* bi = new double[height_sqrt];
-    double* ri = new double[height_sqrt];
+    Matrix<long double> temp(0.0, height, pic_side);
+    Matrix<long double> result(pic_side, pic_side);
+    long double* bi = new long double[height_sqrt];
+    long double* ri = new long double[height_sqrt];
 
     for( size_t i = 0; i < height_sqrt; ++i )
     {
-        bi[i] = 1.0/6.0 + i*(10.-1.0/6.0)/(height_sqrt-1.0);
+        bi[i] = 1.0L/6.0L + i*(10.0L - 1.0L / 6.0L) / ((long double)height_sqrt - 1.0L);
     }
 
     for( size_t i = 0; i < height_sqrt; ++i )
     {
-        ri[i] = 1.0 + i*(pic_side/2.0-1.0)/(height_sqrt-1.0);
+        ri[i] = 1.0L + i*((long double)pic_side / 2.0L - 1.0L) / ((long double)height_sqrt - 1.0L);
     }
 
     size_t row = 0;
@@ -43,12 +43,12 @@ Matrix<double> Spline::Generate( size_t pic_side )
         {
             for( size_t col = 0; col < pic_side/2; ++col )
             {
-                double local_result = std::pow(1.0+std::pow((col+1)/ri[subrow2], 2.0), -3.0*bi[subrow1]);
+                double local_result = std::pow(1.0L + std::pow( ((long double)col + 1.0L) / ri[subrow2], 2.0L), -3.0L * bi[subrow1]);
                 temp[row*pic_side + pic_side/2 + col] = local_result;
                 temp[row*pic_side + pic_side/2 - col - 1] = local_result;
             }
-            double min_value = *std::min_element(&temp[row*pic_side], &temp[(row+1)*pic_side]);
-            double accumulator = 0;
+            long double min_value = *std::min_element(&temp[row*pic_side], &temp[(row+1)*pic_side]);
+            long double accumulator = 0.0L;
             for( size_t i = 0; i < pic_side; ++i )
             {
                 temp[row*pic_side + i] -= min_value;
@@ -66,7 +66,7 @@ Matrix<double> Spline::Generate( size_t pic_side )
     {
         for( size_t j = pic_side/2; j < pic_side; ++j )
         {
-            temp[i*pic_side + j] = 0;
+            temp[i*pic_side + j] = 0.0L;
         }
     }
 
@@ -74,7 +74,7 @@ Matrix<double> Spline::Generate( size_t pic_side )
     {
         for( size_t j = 0; j < pic_side/2; ++j )
         {
-            temp[i*pic_side + j] = 0;
+            temp[i*pic_side + j] = 0.0L;
         }
     }
